Poorva | Extract Command and Query classes
