+++ cli/src/test/resources/gen/array-columns-mapping.expected.txt
@@ -18,12 +18,6 @@ public class FooParamSetter : ParamSetter<FooParams> {
   }
 }
 
-public data class FooResult(
-  public val id: Int,
-  public val colors: Array<String>?,
-  public val details: PGobject?
-)
-
 public class FooRowMapper : RowMapper<FooResult> {
   public override fun map(rs: ResultSet): FooResult = FooResult(
   id = rs.getObject("id") as kotlin.Int,
@@ -39,3 +33,8 @@ public class FooQuery : Query<FooParams, FooResult> {
   public override val paramSetter: ParamSetter<FooParams> = FooParamSetter()
 }
 
+public data class FooResult(
+  public val id: Int,
+  public val colors: Array<String>?,
+  public val details: PGobject?
+)
