+++ cli/src/test/resources/gen/array-contains.expected.txt
@@ -20,11 +20,6 @@ public class FooParamSetter : ParamSetter<FooParams> {
   }
 }
 
-public data class FooResult(
-  public val id: Int,
-  public val colors: Array<String>?
-)
-
 public class FooRowMapper : RowMapper<FooResult> {
   public override fun map(rs: ResultSet): FooResult = FooResult(
   id = rs.getObject("id") as kotlin.Int,
@@ -38,3 +33,8 @@ public class FooQuery : Query<FooParams, FooResult> {
 
   public override val paramSetter: ParamSetter<FooParams> = FooParamSetter()
 }
+
+public data class FooResult(
+  public val id: Int,
+  public val colors: Array<String>?
+)
