+++ cli/src/test/resources/gen/empty-params-class.expected.txt
@@ -16,12 +16,6 @@ public class FooParamSetter : ParamSetter<FooParams> {
   }
 }
 
-public data class FooResult(
-  public val id: Int,
-  public val firstName: String?,
-  public val lastName: String?
-)
-
 public class FooRowMapper : RowMapper<FooResult> {
   public override fun map(rs: ResultSet): FooResult = FooResult(
   id = rs.getObject("id") as kotlin.Int,
@@ -36,3 +30,10 @@ public class FooQuery : Query<FooParams, FooResult> {
 
   public override val paramSetter: ParamSetter<FooParams> = FooParamSetter()
 }
+
+public data class FooResult(
+  public val id: Int,
+  public val firstName: String?,
+  public val lastName: String?
+)
+
