+++ cli/src/test/resources/gen/left-joined-nullable-check.expected.txt
@@ -17,16 +17,6 @@ public class FooParamSetter : ParamSetter<FooParams> {
   }
 }
 
-public data class FooResult(
-  public val id: Int,
-  public val firstName: String?,
-  public val lastName: String?,
-  public val departmentId: Int?,
-  public val name: String?,
-  public val combinationsId: Int?,
-  public val colors: Array<String>?
-)
-
 public class FooRowMapper : RowMapper<FooResult> {
   public override fun map(rs: ResultSet): FooResult = FooResult(
   id = rs.getObject("id") as kotlin.Int,
@@ -50,3 +40,13 @@ public class FooQuery : Query<FooParams, FooResult> {
 
   public override val paramSetter: ParamSetter<FooParams> = FooParamSetter()
 }
+
+public data class FooResult(
+  public val id: Int,
+  public val firstName: String?,
+  public val lastName: String?,
+  public val departmentId: Int?,
+  public val name: String?,
+  public val combinationsId: Int?,
+  public val colors: Array<String>?
+)
