+++ cli/src/test/resources/gen/uuid-column-type-generator.expected.txt
@@ -17,14 +17,6 @@ public class FooParamSetter : ParamSetter<FooParams> {
   }
 }
 
-public data class FooResult(
-  public val id: UUID,
-  public val columnName: String?,
-  public val oldValue: String?,
-  public val newValue: String?,
-  public val capturedAt: OffsetDateTime?
-)
-
 public class FooRowMapper : RowMapper<FooResult> {
   public override fun map(rs: ResultSet): FooResult = FooResult(
   id = rs.getObject("id") as java.util.UUID,
@@ -41,3 +33,11 @@ public class FooQuery : Query<FooParams, FooResult> {
 
   public override val paramSetter: ParamSetter<FooParams> = FooParamSetter()
 }
+
+public data class FooResult(
+  public val id: UUID,
+  public val columnName: String?,
+  public val oldValue: String?,
+  public val newValue: String?,
+  public val capturedAt: OffsetDateTime?
+)
