+++ codegen/src/main/kotlin/norm/model/ColumnModel.kt
@@ -1,8 +1,17 @@
 package norm.model
 
+import com.squareup.kotlinpoet.ARRAY
+import com.squareup.kotlinpoet.ParameterizedTypeName.Companion.parameterizedBy
+import norm.typemapper.DbToKtTypeMapperFactory
+
 data class ColumnModel(
     val fieldName: String,
     val colType: String,
     val colName: String,
     val isNullable: Boolean
-)
+) {
+    fun getTypeName() =
+        if (colType.startsWith("_")) ARRAY.parameterizedBy(DbToKtTypeMapperFactory.getType(colType, false))
+            .copy(nullable = this.isNullable)
+        else DbToKtTypeMapperFactory.getType(colType, isNullable)
+}
